[FLINK-10366] [s3] Consolidate shared classes for S3 in flink-s3-fs-base
Some classes were previously incorrectly in flink-hadoop-fs
